delete some state machine files
